Extract code neutral metadata error
